Final Submission
The drop down menu is populated with the city values.  On change of the
option in the menu, the background changes to match the selected city.
I spiced it up a little and gave the user a prompt after 5 seconds that
they should select an option in the drop down menu to change that
background of the page.